Poject abandoned for now
Basically, there's this big problem of trying to code both a simulation and the real deal in one package. This ended not working as intended (a simple simulation = true/false) because a module called onoff only works on linux, and therefore gives an error on windows and refuses to run. So, by not succeeding to bind the use of that module to the variable simulation = true/false, the project was always only good for one of the two states, and never both at the same time. Plus programming a simulation is kind of sketchy...
Because of this, I decided to stop this project for now. Maybe one day I'll come back to it, though I'll find that unlikely since the arduino project works okayish, there are only a few tweaks I might make to it.